// src/codegen/graphql/graphql.ts
export type Maybe<T> = T | null;
export type InputMaybe<T> = Maybe<T>;
export type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
// ...
export type Scalars = {
  ID: { input: string; output: string; }
  String: { input: string; output: string; }
  Boolean: { input: boolean; output: boolean; }
  Int: { input: number; output: number; }
  Float: { input: number; output: number; }
  Bool: { input: any; output: any; }
  Date: { input: any; output: any; }
  DateTime: { input: any; output: any; }
  /** The `JSON` scalar type represents JSON values as specified by [ECMA-404](https://ecma-international.org/publications-and-standards/standards/ecma-404/). */
  JSON: { input: any; output: any; }
};

export type BlankExperience = IData & _IContent & _IExperience & _IItem & _IPage & {
  __typename?: 'BlankExperience';
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _metadata?: Maybe<IContentMetadata>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
  composition?: Maybe<CompositionStructureNode>;
};
// ...
export type CompositionComponentNode = ICompositionComponentNode & ICompositionNode & {
  __typename?: 'CompositionComponentNode';
  component?: Maybe<_IComponent>;
  displayName?: Maybe<Scalars['String']['output']>;
  displaySettings?: Maybe<Array<Maybe<CompositionDisplaySetting>>>;
  displayTemplateKey?: Maybe<Scalars['String']['output']>;
  key?: Maybe<Scalars['String']['output']>;
  layoutType?: Maybe<Scalars['String']['output']>;
  nodeType?: Maybe<Scalars['String']['output']>;
  type?: Maybe<Scalars['String']['output']>;
};
// ...
export type CompositionStructureNode = ICompositionNode & ICompositionStructureNode & {
  __typename?: 'CompositionStructureNode';
  component?: Maybe<_IComponent>;
  displayName?: Maybe<Scalars['String']['output']>;
  displaySettings?: Maybe<Array<Maybe<CompositionDisplaySetting>>>;
  displayTemplateKey?: Maybe<Scalars['String']['output']>;
  key?: Maybe<Scalars['String']['output']>;
  layoutType?: Maybe<Scalars['String']['output']>;
  nodeType?: Maybe<Scalars['String']['output']>;
  nodes?: Maybe<Array<Maybe<ICompositionNode>>>;
  type?: Maybe<Scalars['String']['output']>;
};
// ...
export type IData = {
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
};
// ...
export type TestimonialCard = IData & _IComponent & _IContent & _IItem & {
  __typename?: 'TestimonialCard';
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _metadata?: Maybe<IContentMetadata>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
  authorName?: Maybe<Scalars['String']['output']>;
  authorRole?: Maybe<Scalars['String']['output']>;
  avatarImage?: Maybe<ContentReference>;
  quoteText?: Maybe<SearchableRichText>;
  rating?: Maybe<Scalars['String']['output']>;
  showAvatar?: Maybe<Scalars['Boolean']['output']>;
};
// ...
export type _IContent = {
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _metadata?: Maybe<IContentMetadata>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
};
// ...
export type _IExperience = {
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _metadata?: Maybe<IContentMetadata>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
  composition?: Maybe<CompositionStructureNode>;
};
// ...
export type _IItem = {
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
};
// ...
export type _IPage = {
  /** @deprecated Use `_link` field instead */
  _children?: Maybe<QueryRef>;
  _deleted?: Maybe<Scalars['Bool']['output']>;
  _fulltext?: Maybe<Array<Maybe<Scalars['String']['output']>>>;
  _id?: Maybe<Scalars['String']['output']>;
  _itemMetadata?: Maybe<_Metadata>;
  _json?: Maybe<Scalars['JSON']['output']>;
  _link?: Maybe<QueryRef>;
  _metadata?: Maybe<IContentMetadata>;
  _modified?: Maybe<Scalars['Date']['output']>;
  _score?: Maybe<Scalars['Float']['output']>;
  _track?: Maybe<Scalars['String']['output']>;
};
// ...
export type GetBlankExperienceByUrlQueryVariables = Exact<{
  url?: InputMaybe<Scalars['String']['input']>;
  path?: InputMaybe<Scalars['String']['input']>;
}>;


export type GetBlankExperienceByUrlQuery = { __typename?: 'Query', BlankExperience?: { __typename?: 'BlankExperienceOutput', total?: number | null, items?: Array<{ __typename?: 'BlankExperience', _metadata?: { __typename?: 'ContentMetadata', displayName?: string | null } | { __typename?: 'InstanceMetadata', displayName?: string | null } | { __typename?: 'ItemMetadata', displayName?: string | null } | { __typename?: 'MediaMetadata', displayName?: string | null } | null, composition?: { __typename?: 'CompositionStructureNode', outline?: Array<(
          { __typename?: 'CompositionComponentNode', displayName?: string | null, nodeType?: string | null }
          & { ' $fragmentRefs'?: { 'CompositionComponentNodeFragment': CompositionComponentNodeFragment } }
        ) | { __typename?: 'CompositionNode', displayName?: string | null, nodeType?: string | null } | (
          { __typename?: 'CompositionStructureNode', displayName?: string | null, nodeType?: string | null }
          & { ' $fragmentRefs'?: { 'CompositionStructureNodeFragmentFragment': CompositionStructureNodeFragmentFragment } }
        ) | null> | null } | null } | null> | null } | null };

export type CompositionComponentNodeFragment = { __typename: 'CompositionComponentNode', type?: string | null, nodeType?: string | null, component?: { __typename: 'BlankSection', _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | { __typename: 'Block', _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | { __typename: 'Hero', _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | { __typename: 'TestimonialCard', quoteText?: { __typename?: 'SearchableRichText', html?: string | null } | null, _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | { __typename: '_Component', _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | { __typename: '_Section', _itemMetadata?: { __typename?: '_Metadata', displayName?: string | null, type?: string | null } | null } | null } & { ' $fragmentName'?: 'CompositionComponentNodeFragment' };

export type CompositionStructureNodeFragmentFragment = { __typename?: 'CompositionStructureNode', nodes?: Array<(
    { __typename?: 'CompositionComponentNode', displayName?: string | null, nodeType?: string | null }
    & { ' $fragmentRefs'?: { 'CompositionComponentNodeFragment': CompositionComponentNodeFragment } }
  ) | { __typename?: 'CompositionNode', displayName?: string | null, nodeType?: string | null } | { __typename?: 'CompositionStructureNode', displayName?: string | null, nodeType?: string | null } | null> | null } & { ' $fragmentName'?: 'CompositionStructureNodeFragmentFragment' };

export class TypedDocumentString<TResult, TVariables>
  extends String
  implements DocumentTypeDecoration<TResult, TVariables>
{
  __apiType?: NonNullable<DocumentTypeDecoration<TResult, TVariables>['__apiType']>;
  private value: string;
  public __meta__?: Record<string, any> | undefined;

  constructor(value: string, __meta__?: Record<string, any> | undefined) {
    super(value);
    this.value = value;
    this.__meta__ = __meta__;
  }

  override toString(): string & DocumentTypeDecoration<TResult, TVariables> {
    return this.value;
  }
}
export const CompositionComponentNodeFragmentDoc = new TypedDocumentString(`
    fragment CompositionComponentNode on CompositionComponentNode {
  __typename
  type
  nodeType
  component {
    __typename
    _itemMetadata {
      displayName
      type
    }
    ... on TestimonialCard {
      __typename
      quoteText {
        html
      }
    }
  }
}
    `, {"fragmentName":"CompositionComponentNode"}) as unknown as TypedDocumentString<CompositionComponentNodeFragment, unknown>;
export const CompositionStructureNodeFragmentFragmentDoc = new TypedDocumentString(`
    fragment CompositionStructureNodeFragment on CompositionStructureNode {
  nodes {
    displayName
    nodeType
    ... on CompositionStructureNode {
      displayName
    }
    ...CompositionComponentNode
  }
}
    fragment CompositionComponentNode on CompositionComponentNode {
  __typename
  type
  nodeType
  component {
    __typename
    _itemMetadata {
      displayName
      type
    }
    ... on TestimonialCard {
      __typename
      quoteText {
        html
      }
    }
  }
}`, {"fragmentName":"CompositionStructureNodeFragment"}) as unknown as TypedDocumentString<CompositionStructureNodeFragmentFragment, unknown>;
export const GetBlankExperienceByUrlDocument = new TypedDocumentString(`
    query GetBlankExperienceByUrl($url: String = "", $path: String = "") {
  BlankExperience(
    where: {_metadata: {url: {base: {eq: $url}, default: {eq: $path}}}}
  ) {
    total
    items {
      _metadata {
        displayName
      }
      composition {
        outline: nodes {
          displayName
          nodeType
          ...CompositionStructureNodeFragment
          ...CompositionComponentNode
        }
      }
    }
  }
}
    fragment CompositionComponentNode on CompositionComponentNode {
  __typename
  type
  nodeType
  component {
    __typename
    _itemMetadata {
      displayName
      type
    }
    ... on TestimonialCard {
      __typename
      quoteText {
        html
      }
    }
  }
}
fragment CompositionStructureNodeFragment on CompositionStructureNode {
  nodes {
    displayName
    nodeType
    ... on CompositionStructureNode {
      displayName
    }
    ...CompositionComponentNode
  }
}`) as unknown as TypedDocumentString<GetBlankExperienceByUrlQuery, GetBlankExperienceByUrlQueryVariables>;